requirements.txt
annotated-types==0.5.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.0
Flask==3.0.0
fonttools==4.43.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
networkx==3.1
numpy==1.26.0
packaging==23.2
Pillow==10.0.1
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
rich==13.6.0
six==1.16.0
typing_extensions==4.8.0
urllib3==2.0.6
Werkzeug==3.0.0